fix: keep link viewer URL stable on hydration
