+++ apps/web/vitest.behavior.browser.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   cacheDir: 'node_modules/.vite-browser-behavior',
-  optimizeDeps: { include: ['mermaid'] },
+  optimizeDeps: { include: ['mermaid', 'react-dom/server'] },
   resolve: { alias: { '~': resolve(import.meta.dirname, 'app') } },
   plugins: [
     tailwindcss(),
